# application/services/custom_services/dbcredential/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: llmchat open api specification
   description: dbcredential API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8012
     description: dbcredential
tags:
- name: dbcredential
paths:
   "/dbcredential/{id}":
      delete:
         tags:
         -  dbcredential
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  dbcredential
         description: GpGetEntityById
         parameters:
         -  in: "path"
            description: GpGetEntityById
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/dbcredential/get/search":
      get:
         tags:
         -  dbcredential
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/dbcredential/get/update":
      put:
         tags:
         -  dbcredential
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/dbcredential"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/dbcredential":
      put:
         tags:
         -  dbcredential
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/dbcredential"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  dbcredential
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  dbcredential
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/dbcredential"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/dbcredential/userid/created_by":
      get:
         tags:
         -  dbcredential
         description: GpGetNounCreatedBy
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      dbcredential:
         type: object
         properties:
            org_ids:
               type: "string"
               description: Description
            db_username:
               type: "string"
               description: Description
            passowrd:
               type: "string"
               description: Description
            host:
               type: "string"
               description: Description
            port:
               type: "string"
               description: Description
            driver:
               type: "string"
               description: Description
            dialect:
               type: "string"
               description: Description
            database_name:
               type: "string"
               description: Description
            enterprise_type:
               type: "string"
               description: Description